+++ _theme_files/readme.txt
@@ -3,3 +3,8 @@ we download from website,
 css,image,html files
 
 then we transfer the html design to componets (by copy past and some small edit)
+
+1. copy html we want( please note the licence when copy)
+2. cut to our componets
+3. ask chatGpt extract out css file to make it clean if needed
+ 3.1 import extracted css file
