+++ src/common/components/Hero.jsx
