+++ src/common/components/Navbar.jsx
@@ -1,5 +1,5 @@
 import { NavLink } from 'react-router-dom';
-import logo from '../assets/images/logo.png';
+import logo from '../../assets/images/logo.png';
 
 const Navbar = () => {
   const linkClass = ({ isActive }) =>
