+++ src/common/pages/HomePage.jsx
@@ -1,0 +1,16 @@
+import Hero from '../components/Hero.jsx';
+import HomeCards from '../components/HomeCards.jsx';
+import UserListings from '../components/UserListings.jsx';
+import ViewAllUsers from '../components/ViewAllUsers.jsx';
+
+const HomePage = () => {
+  return (
+    <>
+      <Hero />
+      <HomeCards />
+      <UserListings isHome={true} />
+      <ViewAllUsers />
+    </>
+  );
+};
+export default HomePage;
