+++ src/common/pages/UserPage.jsx
@@ -6,20 +6,20 @@ import { toast } from 'react-toastify';
 const UserPage = ({ deleteUser }) => {
   const navigate = useNavigate();
   const { id } = useParams();
-  const User = useLoaderData();
+  const user = useLoaderData();
 
-  const onDeleteClick = (UserId) => {
+  const onDeleteClick = (userId) => {
     const confirm = window.confirm(
       'Are you sure you want to delete this listing?'
     );
 
     if (!confirm) return;
 
-    deleteUser(UserId);
+    deleteUser(userId);
 
     toast.success('User deleted successfully');
 
-    navigate('/Users');
+    navigate('/users');
   };
 
   return (
@@ -27,7 +27,7 @@ const UserPage = ({ deleteUser }) => {
       <section>
         <div className='container m-auto py-6 px-6'>
           <Link
-            to='/Users'
+            to='/users'
             className='text-indigo-500 hover:text-indigo-600 flex items-center'
           >
             <FaArrowLeft className='mr-2' /> Back to User Listings
@@ -40,11 +40,11 @@ const UserPage = ({ deleteUser }) => {
           <div className='grid grid-cols-1 md:grid-cols-70/30 w-full gap-6'>
             <main>
               <div className='bg-white p-6 rounded-lg shadow-md text-center md:text-left'>
-                <div className='text-gray-500 mb-4'>{User.type}</div>
-                <h1 className='text-3xl font-bold mb-4'>{User.title}</h1>
+                <div className='text-gray-500 mb-4'>{user.type}</div>
+                <h1 className='text-3xl font-bold mb-4'>{user.title}</h1>
                 <div className='text-gray-500 mb-4 flex align-middle justify-center md:justify-start'>
                   <FaMapMarker className='text-orange-700 mr-1' />
-                  <p className='text-orange-700'>{User.location}</p>
+                  <p className='text-orange-700'>{user.location}</p>
                 </div>
               </div>
 
@@ -53,13 +53,13 @@ const UserPage = ({ deleteUser }) => {
                   User Description
                 </h3>
 
-                <p className='mb-4'>{User.description}</p>
+                <p className='mb-4'>{user.description}</p>
 
                 <h3 className='text-indigo-800 text-lg font-bold mb-2'>
                   Salary
                 </h3>
 
-                <p className='mb-4'>{User.salary} / Year</p>
+                <p className='mb-4'>{user.salary} / Year</p>
               </div>
             </main>
 
@@ -68,36 +68,36 @@ const UserPage = ({ deleteUser }) => {
               <div className='bg-white p-6 rounded-lg shadow-md'>
                 <h3 className='text-xl font-bold mb-6'>Company Info</h3>
 
-                <h2 className='text-2xl'>{User.company.name}</h2>
+                <h2 className='text-2xl'>{user.company.name}</h2>
 
-                <p className='my-2'>{User.company.description}</p>
+                <p className='my-2'>{user.company.description}</p>
 
                 <hr className='my-4' />
 
                 <h3 className='text-xl'>Contact Email:</h3>
 
                 <p className='my-2 bg-indigo-100 p-2 font-bold'>
-                  {User.company.contactEmail}
+                  {user.company.contactEmail}
                 </p>
 
                 <h3 className='text-xl'>Contact Phone:</h3>
 
                 <p className='my-2 bg-indigo-100 p-2 font-bold'>
                   {' '}
-                  {User.company.contactPhone}
+                  {user.company.contactPhone}
                 </p>
               </div>
 
               <div className='bg-white p-6 rounded-lg shadow-md mt-6'>
                 <h3 className='text-xl font-bold mb-6'>Manage User</h3>
                 <Link
-                  to={`/edit-User/${User.id}`}
+                  to={`/edit-user/${user.id}`}
                   className='bg-indigo-500 hover:bg-indigo-600 text-white text-center font-bold py-2 px-4 rounded-full w-full focus:outline-none focus:shadow-outline mt-4 block'
                 >
                   Edit User
                 </Link>
                 <button
-                  onClick={() => onDeleteClick(User.id)}
+                  onClick={() => onDeleteClick(user.id)}
                   className='bg-red-500 hover:bg-red-600 text-white font-bold py-2 px-4 rounded-full w-full focus:outline-none focus:shadow-outline mt-4 block'
                 >
                   Delete User
@@ -111,10 +111,10 @@ const UserPage = ({ deleteUser }) => {
   );
 };
 
-const UserLoader = async ({ params }) => {
-  const res = await fetch(`/api/Users/${params.id}`);
+const userLoader = async ({ params }) => {
+  const res = await fetch(`/api/users/${params.id}`);
   const data = await res.json();
   return data;
 };
 
-export { UserPage as default, UserLoader };
+export { UserPage as default, userLoader };
