+++ src/layouts/MainLayout.jsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import Navbar from '../components/Navbar';
+import Navbar from '../common/components/Navbar.jsx';
 
 const MainLayout = () => {
   return (
