+++ src/page1/readme.txt
@@ -1,0 +1,8 @@
+this is a template folder to structure our code,
+you can create folder like userinfo, useretc..
+instead of put all our pages, componets together,
+
+better create own folder to store pages and componets,
+dont messup others work.
+
+can refer to register folder as example.
