--- src/pages/HomePage.jsx
@@ -1,16 +1,0 @@
-import Hero from '../components/Hero';
-import HomeCards from '../components/HomeCards';
-import UserListings from '../components/UserListings';
-import ViewAllUsers from '../components/ViewAllUsers';
-
-const HomePage = () => {
-  return (
-    <>
-      <Hero />
-      <HomeCards />
-      <UserListings isHome={true} />
-      <ViewAllUsers />
-    </>
-  );
-};
-export default HomePage;
